+++ Runtime/Scripts/Utilities/Animations/SimpleSpin.cs
@@ -1,9 +1,14 @@
 using UnityEngine;
+using Utilities.Enums;
+using Utilities.Extensions;
 
 namespace Utilities.Animations
 {
     public class SimpleSpin : MonoBehaviour
     {
+        [SerializeField]
+        private SPACE space = SPACE.WORLD;
+        
         public bool reverse;
         [SerializeField]
         private Vector3 spin;
@@ -15,7 +20,7 @@ namespace Utilities.Animations
 
             currentRotation *= Quaternion.Euler(spin * (Time.deltaTime * (reverse ? -1 : 1)));
 
-            transform.rotation = currentRotation;
+            transform.SetRotation(space, currentRotation);
         }
     }
 }